Fix tree access in git.Repository.read_text on Windows
The function git.Repository.read_text needs to access a blob in the tree
associated with a commit. To do so, it uses operator.truediv (`/`) with
commit.tree and the string representation of the given path, relative to
the repository.

Using the string representation of the path does not work on Windows,
because `str(path)` returns a backslash-separated path on this platform.
When accessing trees with libgit2, paths must always be separated by
slashes.

Use operator.truediv (`/`) to append all the components of the path to
commit.tree, one by one. This ensures a valid path representation for
tree access with libgit2.