Allow passing commits as positional arguments (#39)
* Add commits argument to retrocookie

* Add core.get_commits to identify commits for cherry picking

* Add positional arguments <commits>

* Adapt tests for command-line interface

* Add test case for `commits` parameter to retrocookie

* Minor cleanup in tests.helpers.commit

* Do not invoke git-rev-list with range and single commit

In unit tests, this resulted in returning [] for "HEAD". In interactive
use with a real repository, "HEAD" was traversed despite --no-walk,
likely due to also passing a range expression. In both cases, the first
argument passed to git rev-list was `master..master`, which correctly
resulted in an empty list.

* Read from object database in tests.helpers.read, not worktree

This spares us the trouble of updating the worktree in tests. (Fixes
test_commit with the latest changes on this branch.)

* Fix tree access in git.Repository.read_text on Windows

The function git.Repository.read_text needs to access a blob in the tree
associated with a commit. To do so, it uses operator.truediv (`/`) with
commit.tree and the string representation of the given path, relative to
the repository.

Using the string representation of the path does not work on Windows,
because `str(path)` returns a backslash-separated path on this platform.
When accessing trees with libgit2, paths must always be separated by
slashes.

Use operator.truediv (`/`) to append all the components of the path to
commit.tree, one by one. This ensures a valid path representation for
tree access with libgit2.