flake8: Ignore multiple spaces after ',' (E241) in test_git
This test module uses manually formatted tables in `pytest.mark.parametrize`,
surrounded by `# fmt: [on|off]` markers. Each line in the table generates a
warning. As flake8 does not support ignoring a warning for a region of text, we
use a per-file ignore instead of applying `# noqa: E241` to each line.