Clean up prompts in doctests
The PS1 prompt must be used when a doctest contains multiple statements.
Continuations may be preceded by the PS2 prompt---this is mandatory in the
doctest module, and optional in xdoctest.

A doctest with a single PS1 prompt uses the `compile` builtin in single mode,
and a `SyntaxError: invalid syntax` is raised. This is true of both doctest and
xdoctest.