📚 Use furo Sphinx theme for documentation (#411)
* ➕ [poetry] Add furo >=2021.11.12 to development dependencies

* 👷 [nox] Install furo in docs and docs-build sessions

* ➕ [docs] Add furo==2021.11.12 to requirements

* 🔧 [docs] Use furo Sphinx theme

* 📚 [docs] Remove table of contents from API Reference

The table of contents duplicates furo's right sidebar.

* 📚 [docs] Generate API documentation for package not `__main__`

`reference.rst` inserts the API documentation for the `__main__` module. That
makes little sense, as it only contains the Click entry point, whose docstring
is not intended for consumption by Sphinx.

Instead, insert API documentation for the package. This generates documentation
for anything in `__init__`. On project generation, this is only be the package
docstring, which is fine.

* ➖ [docs] Remove sphinx-rtd-theme from requirements

* 👷 [nox] Do not install sphinx-rtd-theme in docs and docs-build sessions

* ➖ [poetry] Remove sphinx-rtd-theme from development dependencies

Retrocookie-Original-Commit: cjolowicz/cookiecutter-hypermodern-python-instance@272d565